Add files via upload
After the submission of the report on May 19, further analysis was done, more data was collected and a new model was successfully constructed with a 79% accuracy score in student performance prediction. This was the best result obtained, therefore the paper, dataset and codes were updated and uploaded to GitHub. 


You can find the old submission in this repository, however the new results are summarized in the documents which have names starting with "Final update".